% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalFCnet.R
\name{evalFCnet}
\alias{evalFCnet}
\title{Evaluate model's prediction against the observed true values}
\usage{
evalFCnet(model)
}
\arguments{
\item{model}{an object returned from a call to cv_FCnet}
}
\value{
A dataframe with model's goodness of fit statistics.
}
\description{
Evaluate model's prediction against the observed true values
}
