% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_CVparsFCnet.R
\name{get_CVparsFCnet}
\alias{get_CVparsFCnet}
\title{Internal function for retrieving crossvalidated parameters of cv_FCnet}
\usage{
get_CVparsFCnet(fit)
}
\arguments{
\item{fit}{The default color palette for plotting matrices (a vector of colors).}
}
\value{
The crossvalidated parameters with the associated error
}
\description{
Internal function to retrieve the alpha and lambda parameters
resulting in the minimum crossvalidation error.
}
