% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optionsFCnet.R
\name{optionsFCnet}
\alias{optionsFCnet}
\title{Set or get options for FCnet}
\usage{
optionsFCnet(...)
}
\arguments{
\item{cv.type.measure}{The measure to minimize in crossvalidation inner loops.
Differently from \code{glmnetUtils::cva.glmnet()} the deafult is the mean absolute error.}

\item{intercept}{whether to fit (TRUE) or not (FALSE) an intercept to the model.}

\item{standardize}{Whether x must be standardized. Differently from
\code{glmnet::glmnet()} the default is FALSE as we assume predictors are already either
summarised with PCA or ICA (and therefore scaled) or drawn from normalized FC matrices.}

\item{colorPaletteDefault}{The default color palette for plotting matrices (a vector of colors).}
}
\description{
Set or get options for FCnet
}
