% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_volume.R
\name{plot_volume}
\alias{plot_volume}
\title{plot a brain Volume}
\usage{
plot_volume(
  volume,
  x = 1,
  y = 1,
  z = 1,
  limit = NULL,
  col.y = c("darkslateblue", "black", "firebrick4"),
  colorNuances = optionsFCnet("colorNuances")
)
}
\arguments{
\item{volume}{A 3_d array to depict.}

\item{x}{Coordinates at which the orthostatic view is displayed.}

\item{y}{Coordinates at which the orthostatic view is displayed.}

\item{z}{Coordinates at which the orthostatic view is displayed.}

\item{limit}{Range of values to depict. If NULL, the default,
simmetry is preserved around the largest value in absolute value.}

\item{col.y}{The color palette used to plot values (i.e. a vector of colors). The vector is often divergent, es. \code{c("red", "white", "blue")}.
Defaults to \code{optionsFCnet("colorPaletteDefault")}.}

\item{colorNuances}{Number of nuances along the provided col.y.}
}
\description{
This function reads a 3-D array and returns a very basic plot
using \code{neurobase::ortho2()}. This can be, for example, a back-projected volume
or the mean volume returned from the reduce_features object.
}
