% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFCnet.R
\name{plotFCnet}
\alias{plotFCnet}
\title{plot a FCnet model}
\usage{
plotFCnet(
  model,
  output = c("predictions", "coefficients"),
  style = c("boxplot", "violin"),
  plot_labels = T,
  subset_coeffs = NULL
)
}
\arguments{
\item{model}{The FCnet model to depict.}

\item{output}{What to depict: model's predictions or models' coefficients. Coefficients
are assumed to arise from nested models, thus they will depicted with - more or less pronounced -
variability with the geometries chosen with the style parameter.}

\item{style}{For coefficients, how to represent variability (e.g. boxplots or violins).}

\item{plot_labels}{if \code{TRUE} annotates the row names of x as predictions.}

\item{subset_coeffs}{To avoid padding, sometimes the user may want to subset the coefficients to report.
This can be done by passing a vector to this parameter. Note that the intercept counts as
a coefficient (of 0) even if omitted in the call.}
}
\value{
A \code{ggplot2} object which can be customized further.
}
\description{
This function accepts a FCnet model - such as one retuned by
\code{FCnetLOO()} - and returns a plot for either the coefficients or the
model's prediction.
}
