% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_indices.R
\name{reorder_network_list}
\alias{reorder_network_list}
\title{Reorder network list with Schaefer networks first, others at end}
\usage{
reorder_network_list(indices)
}
\arguments{
\item{indices}{Named list of ROI indices}
}
\value{
Reordered list
}
\description{
Reorder network list with Schaefer networks first, others at end
}
\keyword{internal}
