% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_FCnet.R
\name{run_FCnet}
\alias{run_FCnet}
\title{FCnet (experimental) shiny Graphic User Interface}
\usage{
run_FCnet()
}
\description{
This function launches an experimental GUI for FCnet
implementing a slightly simplified pipeline.
}
