% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_coefficients.R
\name{select_coefficients}
\alias{select_coefficients}
\title{Selection of coefficients based on the outcome of permutateLOO}
\usage{
select_coefficients(coeffs, permutated_coeffs, threshold = 0.05)
}
\arguments{
\item{coeffs}{Coefficients can either be
passed as a list (as one created, for example, by \code{FCnetLOO()}), or as a data.frame (which must, however,
have the same structure as the one returned by \code{FCnetLOO()}).
This function finds consensus coefficients first trough the function specified
by \code{optionsFCnet("consensus_function")},which by default is the median.}

\item{permutated_coeffs}{A list, such as an object created by
\code{permutateLOO()} (with the parameter \code{return_coeffs} set to \code{TRUE}),
or a data.frame (which must, however, have the same structure
as the one returned by \code{permutateLOO()})}

\item{threshold}{Significance threshold against which evaluate the coefficients.
It is assumed to be two-tailed, and defaults to 0.05.}
}
\value{
A list with two entries: coeffs, i.e. the values of coefficients passing the
statistical threshold, and mask, reporting which coefficients are retained.
}
\description{
This function assesses the original coefficients against the ones obtained
from permutated samples as returned by \code{permutateLOO()} (with the parameter
\code{return_coeffs} set to \code{TRUE}). It returns a vector of weights as well as a
dichotomic mask reflecting which values pass a given threshold.
}
