% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optionsFCnet.R
\name{optionsFCnet}
\alias{optionsFCnet}
\title{Set or get options for FCnet}
\usage{
optionsFCnet(...)
}
\arguments{
\item{family}{Defaults to "gaussian." Experimental support for "binomial" on the way.}

\item{cv.type.measure}{The measure to minimize in crossvalidation inner loops.
The deafult is the mean squared error.}

\item{cv.criterion}{Whether crossvalidated models are chosen on the basis of
their error or - relevant when different number of components are compared -
AIC or BIC criteria.}

\item{intercept}{whether to fit (TRUE) or not (FALSE) an intercept to the model.
Useful if y is not scaled and centered.}

\item{standardize}{Whether x must be standardized internally to \code{glmnet::glmnet()}.}

\item{whichLambda}{During crossvalidation, either selects "lambda.min" - which
chooses the value of lambda which best minimizes the mean crossvalidation error - or
"lambda.1se" - which "which gives the most regularized model such that error is
within one standard error of the minimum". The latter would be preferable
for stability.}

\item{thresh}{Threshold for glmnet to stop converging to the solution.
This parameter can be adapted (e.g. for speed / accuracy tradeoff).}

\item{consensus_function}{which function is used to create consensus between
coefficients from different models.}

\item{nested}{Whether crossvalidation is to be performed in a nested fashion.}

\item{colorPaletteDefault}{The default color palette for plotting matrices (a vector of colors).}

\item{colorNuances}{Number of nuances along the provided colorPalette.}
}
\description{
Set or get options for FCnet
}
