% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optionsFCnet.R
\name{optionsFCnet}
\alias{optionsFCnet}
\title{Set or get options for FCnet}
\usage{
optionsFCnet(...)
}
\arguments{
\item{colorPaletteDefault}{The default color palette for plotting matrices (a vector of colors).}
}
\description{
Set or get options for FCnet
}
