% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optionsFCnet.R
\name{optionsFCnet}
\alias{optionsFCnet}
\title{Set or get options for FCnet}
\usage{
optionsFCnet(...)
}
\arguments{
\item{cv.type.measure}{The measure to minimize in crossvalidation inner loops.
The deafult is the mean squared error.}

\item{intercept}{whether to fit (TRUE) or not (FALSE) an intercept to the model.
Useful if y is not scaled and centered.}

\item{standardize}{Whether x must be standardized internally to \code{glmnet::glmnet()}.}

\item{whichLambda}{During crossvalidation, either selects "lambda.min" - which
chooses the value of lambda which best minimizes the mean crossvalidation error - or
"lambda.1se" - which "which gives the most regularized model such that error is
within one standard error of the minimum". The latter would be preferable
for stability.
@param consensus_function which function is used to create consensus between
coefficients from different models.}

\item{colorPaletteDefault}{The default color palette for plotting matrices (a vector of colors).}

\item{colorNuances}{Number of nuances along the provided colorPalette.}
}
\description{
Set or get options for FCnet
}
